requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
amqp==5.1.1
attrs==21.2.0
bs4==0.0.1
celery==5.2.7
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.8
configparser==5.1.0
crayons==0.4.0
cryptography==42.0.5
distlib==0.3.6
et-xmlfile==1.1.0
execnet==1.9.0
Faker==9.8.3
greenlet==3.0.3
h11==0.12.0
h2==4.1.0
html5lib==1.1
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
install==1.3.5
itsdangerous==2.1.2
jeepney==0.8.0
kaitaistruct==0.9
msgpack==1.0.3
openpyxl==3.0.9
outcome==1.1.0
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.1
platformdirs==2.4.0
playwright==1.41.2
pluggy==1.4.0
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pydivert==2.1.0
pyee==11.0.1
pylev==1.4.0
pyOpenSSL==24.0.0
pyparsing==3.0.6
pyrsistent==0.20.0
pytest==8.0.0
pytest-rerunfailures==10.2
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.20.0
pytz==2022.2.1
pywin32-ctypes==0.2.0
requests==2.26.0
requests-toolbelt==0.9.1
shellingham==1.4.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tomlkit==0.7.2
tqdm==4.64.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.7
urllib3-secure-extra==0.1.0
virtualenv==20.10.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
wsproto==1.0.0
zipp==3.8.0
